+++ tests/unit/client/feedback/dataset/test_base.py
@@ -29,7 +29,11 @@ from argilla.client.feedback.schemas.records import FeedbackRecord
 from argilla.client.sdk.v1.datasets.models import FeedbackResponseStatusFilter
 
 if TYPE_CHECKING:
-    from argilla.client.feedback.schemas.types import AllowedFieldTypes, AllowedQuestionTypes
+    from argilla.client.feedback.schemas.types import (
+        AllowedFieldTypes,
+        AllowedMetadataPropertyTypes,
+        AllowedQuestionTypes,
+    )
 
 
 class TestFeedbackDataset(FeedbackDatasetBase):
@@ -260,3 +264,24 @@ def test__parse_and_validate_records_validation_error(
     )
     with pytest.raises(exception_cls, match=exception_msg):
         dataset._parse_and_validate_records(record)
+
+
+@pytest.mark.parametrize(
+    "metadata_property",
+    (
+        TermsMetadataProperty(name="terms-metadata-diff-name", values=["a", "b", "c"]),
+        IntegerMetadataProperty(name="int-metadata-diff-name", min=0, max=10),
+        FloatMetadataProperty(name="float-metadata-diff-name", min=0.0, max=10.0),
+    ),
+)
+def test__unique_metadata_property(metadata_property: "AllowedMetadataPropertyTypes") -> None:
+    dataset = TestFeedbackDataset(
+        fields=[TextField(name="required-field", required=True), TextField(name="optional-field", required=False)],
+        questions=[TextQuestion(name="question", required=True)],
+        metadata_properties=[
+            TermsMetadataProperty(name="terms-metadata", values=["a", "b", "c"]),
+            IntegerMetadataProperty(name="int-metadata", min=0, max=10),
+            FloatMetadataProperty(name="float-metadata", min=0.0, max=10.0),
+        ],
+    )
+    dataset._unique_metadata_property(metadata_property)